Combine pinpoint.py into lithium_helpers.py
