+++ src/funfuzz/js/loop.py
@@ -18,7 +18,6 @@ from optparse import OptionParser  # pylint: disable=deprecated-module
 
 from . import compare_jit
 from . import js_interesting
-from . import pinpoint
 from . import shell_flags
 from ..util import create_collector
 from ..util import file_manipulation
@@ -46,7 +45,7 @@ def parseOpts(args):  # pylint: disable=invalid-name,missing-docstring,missing-r
     parser.add_option("--build",
                       action="store", dest="build_options_str",
                       help="The build options, for bisection",
-                      default=None)  # if you run loop directly without --build, pinpoint will try to guess
+                      default=None)  # if you run loop directly w/o --build, lithium_helpers.pinpoint will try to guess
     parser.add_option("--valgrind",
                       action="store_true", dest="valgrind",
                       default=False,
@@ -187,7 +186,7 @@ def many_timed_runs(targetTime, wtmpDir, args, collector):  # pylint: disable=in
             itest.append("--minlevel=" + str(res.lev))
             itest.append("--timeout=" + str(options.timeout))
             itest.append(options.knownPath)
-            (lithResult, _lithDetails, autoBisectLog) = pinpoint.pinpoint(  # pylint: disable=invalid-name
+            (lithResult, _lithDetails, autoBisectLog) = lithium_helpers.pinpoint(  # pylint: disable=invalid-name
                 itest, logPrefix, options.jsEngine, engineFlags, filenameToReduce, options.repo,
                 options.build_options_str, targetTime, res.lev)
